Moving fixtures to its own file
